adds code to pass first photograph test
