+++ lib/photograph.rb
@@ -1,0 +1,6 @@
+class Photograph
+  def initialize(data)
+    
+  end
+
+end
